EONEXT-71 - Created translations as separate component.
