--- src/components/search-bar/header-styles-extended.scss
@@ -1,79 +1,0 @@
-.header {
-  &__menu-first {
-    display: flex;
-    flex-direction: column-reverse;
-  }
-
-  &__menu-navigation-buttons {
-    display: flex;
-    justify-content: flex-end;
-  }
-
-  &__select-wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: flex-start;
-    position: relative;
-    width: 110px;
-  }
-
-  &__select {
-    background: transparent;
-    border: none;
-    outline: none;
-    font-family: "Noto Sans", sans-serif;
-    font-style: normal;
-    font-weight: 400;
-    font-size: 12px;
-    cursor: initial;
-
-    -webkit-appearance: none;
-    -moz-appearance: none;
-    appearance: none;
-    .goog-te-gadget .goog-te-combo {
-      margin-top: 0;
-    }
-    &::-ms-expand {
-      display: none;
-    }
-    span {
-      display: none;
-    }
-    .skiptranslate {
-      color: transparent;
-    }
-  }
-
-  .goog-te-combo {
-    background: transparent;
-    border: none;
-    outline: none;
-    padding-top: 47px;
-    cursor: pointer;
-    text-align: center;
-    text-transform: capitalize;
-
-    -webkit-appearance: none;
-    -moz-appearance: none;
-    appearance: none;
-    &::-ms-expand {
-      display: none;
-    }
-
-    option {
-      font-family: "Noto Sans", sans-serif;
-      font-style: normal;
-      font-weight: 400;
-      font-size: 12px;
-    }
-  }
-
-  &__select-icon {
-    width: 24px;
-    height: 24px;
-    position: absolute;
-    top: 22px;
-    z-index: -1;
-  }
-}
